Organized code a little better, seperated out game code from non game code.
