+++ gameCode/bugMoveLogic/bugLogic/ant.logic.js
