+++ gameCode/bugMoveLogic/bugLogic/beetle.logic.js
