+++ gameCode/bugMoveLogic/bugLogic/hopper.logic.js
