+++ gameCode/bugMoveLogic/bugLogic/queenBee.logic.js
