+++ gameCode/bugMoveLogic/bugLogic/spider.logic.js
