+++ gameCode/bugMoveLogic/continuity.util.js
