+++ gameCode/bugMoveLogic/logic.util.js
@@ -1,6 +1,3 @@
-import Board from "../gameObjects/board.js"
-import Bug from "../gameObjects/bugs.js"
-import CellState from "../gameObjects/cellState.js"
 import { translateRefCoordToArrayCoord } from "../utils/coordinateTranslate.util.js";
 
 export function checkCrescentBreak(move, board) {
