+++ gameCode/gameObjects/bugs.js
