+++ gameCode/gameObjects/cellState.js
